backend/requirements.txt
fastapi
uvicorn
ultralytics
numpy
Pillow
opencv-python-headless
matplotlib
torch
torchaudio
torchvision